Update 3.2 pre-release headers
Minor tweaks to two function signatures and a new error code
appeared.

This should be very close to the final headers. I do not expect
any other breaking change at this point.

Also drop a structure that is not defined in the spec and I
believe will be dropped from the Oasis vesion as well eventually.

Signed-off-by: Simo Sorce <simo@redhat.com>